+++ src/components/AnswerItem.vue
@@ -1,5 +1,7 @@
 <template>
-  <button @click.prevent="$emit('button-clicked',props.answer)">{{ props.answer }}</button>
+  <button @click.prevent="$emit('button-clicked', props.answer)">
+    {{ props.answer }}
+  </button>
 </template>
 
 <script setup>
@@ -9,6 +11,4 @@ const props = defineProps({
     required: true,
   },
 });
-
-
 </script>
