+++ src/views/LandingView.vue
@@ -1,11 +1,11 @@
 <template>
-  <main id="landing-page-box">
+  <main>
     <img
       id="logo-img"
       src="../assets/images/logo.png"
       alt="Paefax logo globe with flags"
     />
-    <h1 id="landing-page-title">Paefax Language Quiz</h1>
+    <h1>Paefax Language Quiz</h1>
     <RouterLink to="/language">
       <button>Get Started</button>
     </RouterLink>
@@ -15,7 +15,7 @@
 <script setup></script>
 
 <style scoped>
-#landing-page-box {
+main {
   display: flex;
   flex-direction: column;
   gap: 10px;
@@ -26,9 +26,5 @@
 
 #logo-img {
   max-width: 80%;
-}
-
-#landing-page-title {
-  font-size: 1.5em;
 }
 </style>
